:wheelchair: more logging and improve code coherence
